WIP - Created RoadAddressChangeInfoMapper in order to deal with the roadChanges  type 1 and 2
